+++ main/pomaquet.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	"github.com/padilo/pomaquet/pkg/tui"
+	"github.com/padilo/pomaquet/pkg/core/tui"
 )
 
 func main() {
