--- pkg/app/pomodoro/model/messages.go
@@ -1,14 +1,0 @@
-package model
-
-type MsgError struct {
-	Err error
-}
-
-type MsgPomodoroStarted struct {
-}
-
-type MsgPomodoroFinished struct {
-}
-
-type MsgPomodoroCancelled struct {
-}
