--- pkg/app/task/task.go
@@ -1,8 +1,0 @@
-package task
-
-type Task struct {
-	Done  bool
-	Title string
-}
-
-type TaskList []Task
