+++ pkg/core/infrastructure/config_dir.go
@@ -1,0 +1,22 @@
+package infrastructure
+
+import (
+	"log"
+	"os"
+	"os/user"
+	"path"
+)
+
+func ConfigDir() string {
+	u, err := user.Current()
+	if err != nil {
+		panic(err)
+	}
+	dir := path.Join(u.HomeDir, "/.pomaquet")
+
+	err = os.MkdirAll(dir, 0700)
+	if err != nil {
+		log.Fatalf("Couldn't create %v directory\n", dir)
+	}
+	return dir
+}
