+++ pkg/core/tui/run.go
@@ -1,0 +1,23 @@
+package tui
+
+import (
+	"fmt"
+	"os"
+
+	tea "github.com/charmbracelet/bubbletea"
+	"github.com/padilo/pomaquet/pkg/core/tui/window"
+)
+
+func Run() {
+	p := tea.NewProgram(
+		window.NewModel(),
+		tea.WithAltScreen(),
+		tea.WithMouseCellMotion(),
+	)
+
+	if err := p.Start(); err != nil {
+		fmt.Println("could not run program:", err)
+		os.Exit(1)
+	}
+
+}
