+++ pkg/pomodoro/domain/messages.go
@@ -1,0 +1,14 @@
+package domain
+
+type MsgError struct {
+	Err error
+}
+
+type MsgPomodoroStarted struct {
+}
+
+type MsgPomodoroFinished struct {
+}
+
+type MsgPomodoroCancelled struct {
+}
