+++ pkg/settings/domain/repository.go
@@ -1,0 +1,6 @@
+package domain
+
+type SettingsRepository interface {
+	Save(settings Settings)
+	Get() Settings
+}
