--- pkg/storage/interface.go
@@ -1,10 +1,0 @@
-package storage
-
-import (
-	"github.com/padilo/pomaquet/pkg/app/pomodoro/model"
-)
-
-type SettingsStorage interface {
-	Save(settings model.Settings)
-	Get() model.Settings
-}
