+++ pkg/task/domain/task.go
@@ -1,0 +1,8 @@
+package domain
+
+type Task struct {
+	Done  bool
+	Title string
+}
+
+type TaskList []Task
