--- pkg/tui/run.go
@@ -1,23 +1,0 @@
-package tui
-
-import (
-	"fmt"
-	"os"
-
-	tea "github.com/charmbracelet/bubbletea"
-	"github.com/padilo/pomaquet/pkg/tui/window"
-)
-
-func Run() {
-	p := tea.NewProgram(
-		window.NewModel(),
-		tea.WithAltScreen(),
-		tea.WithMouseCellMotion(),
-	)
-
-	if err := p.Start(); err != nil {
-		fmt.Println("could not run program:", err)
-		os.Exit(1)
-	}
-
-}
